# swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión Empresarial
  description: |
    API para gestionar usuarios, compañías, clientes, proyectos y albaranes en un sistema de gestión empresarial.
    Permite administrar el ciclo completo de proyectos, desde la creación hasta la generación de albaranes.
  version: 1.0.0
  contact:
    email: admin@ejemplo.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
          description: ID único del usuario
        email:
          type: string
          format: email
          description: Correo electrónico (único)
        nombre:
          type: string
          description: Nombre del usuario
        apellidos:
          type: string
          description: Apellidos del usuario
        nif:
          type: string
          description: Número de Identificación Fiscal
        direccion:
          type: string
          description: Dirección física del usuario
        isValidated:
          type: boolean
          description: Indica si la cuenta está validada
      required:
        - email
        - nombre
        - apellidos
        - nif
        - direccion
...
  responses:
    Unauthorized:
      description: No autorizado - es necesario autenticarse
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: No autorizado
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Recurso no encontrado
    BadRequest:
      description: Error en la solicitud
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Error en los datos enviados
security:
  - cookieAuth: []
paths:
  /register:
    post:
      tags:
        - Usuarios
      summary: Registrar un nuevo usuario
      description: Crea una nueva cuenta de usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                nombre:
                  type: string
                apellidos:
                  type: string
                nif:
                  type: string
                direccion:
                  type: string
              required:
                - email
                - password
                - nombre
                - apellidos
                - nif
                - direccion
      responses:
        '201':
          description: Usuario registrado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
                  validationCode:
                    type: string
                    description: Código para validar la cuenta
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/user/validation:
    put:
      tags:
        - Usuarios
      summary: Validar cuenta de usuario
      description: Valida la cuenta de usuario con un código
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
              required:
                - email
                - code
      responses:
        '200':
          description: Cuenta validada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/user/resend-code:
    post:
      tags:
        - Usuarios
      summary: Reenviar código de validación
      description: Genera y envía un nuevo código de validación al usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Código reenviado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  validationCode:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /login:
    post:
      tags:
        - Usuarios
      summary: Iniciar sesión
      description: Inicia sesión con email y contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/user/change-password:
    post:
      tags:
        - Usuarios
      summary: Cambiar contraseña
      description: Cambia la contraseña del usuario autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
              required:
                - currentPassword
                - newPassword
      responses:
        '200':
          description: Contraseña cambiada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/info:
    get:
      tags:
        - Usuarios
      summary: Obtener información del usuario
      description: Devuelve la información del usuario autenticado
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /companies/create:
    post:
      tags:
        - Compañías
      summary: Crear nueva compañía
      description: Crea una nueva compañía con el usuario como jefe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emailJefe:
                  type: string
                  format: email
                nif:
                  type: string
                nombre:
                  type: string
                miembros:
                  type: array
                  items:
                    type: string
                    format: email
              required:
                - emailJefe
                - nif
                - nombre
      responses:
        '201':
          description: Compañía creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/companies:
    get:
      tags:
        - Compañías
      summary: Listar compañías del usuario
      description: Devuelve las compañías donde el usuario es jefe o miembro
      responses:
        '200':
          description: Lista de compañías
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error del servidor
  /api/clients:
    get:
      tags:
        - Clientes
      summary: Listar todos los clientes
      description: Devuelve todos los clientes activos
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error del servidor
    post:
      tags:
        - Clientes
      summary: Crear nuevo cliente
      description: Crea un nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                apellidos:
                  type: string
                email:
                  type: string
                  format: email
                telefono:
                  type: string
                nif:
                  type: string
                direccion:
                  type: string
                companiaId:
                  type: string
                  format: ObjectId
              required:
                - nombre
                - apellidos
                - email
                - telefono
                - nif
                - direccion
      responses:
        '201':
          description: Cliente creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/clients/me:
    get:
      tags:
        - Clientes
      summary: Listar clientes creados por el usuario
      description: Devuelve los clientes creados por el usuario autenticado
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error del servidor
...
  /api/projects/me:
    get:
      tags:
        - Proyectos
      summary: Listar proyectos creados por el usuario
      description: Devuelve los proyectos creados por el usuario autenticado
      responses:
        '200':
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error del servidor
  /api/projects/client/{clientId}:
    get:
      tags:
        - Proyectos
      summary: Listar proyectos por cliente
      description: Devuelve los proyectos asociados a un cliente específico
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error del servidor
...
  /api/albaranes/pdf/{id}:
    get:
      tags:
        - Albaranes
      summary: Generar PDF del albarán
      description: Genera y devuelve un archivo PDF del albarán (solo si está firmado)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: ObjectId
      responses: